[GEOT-3240] Shapefile data store will fail when the .dbf file size exceeds 2GB
git-svn-id: http://svn.osgeo.org/geotools/trunk@36014 e5c1c795-43da-0310-a71f-fac65c449510

+++ modules/plugin/shapefile/src/main/java/org/geotools/data/shapefile/dbf/IndexedDbaseFileReader.java
@@ -67,11 +67,11 @@ public class IndexedDbaseFileReader extends DbaseFileReader {
             UnsupportedOperationException {
 
         if (this.randomAccessEnabled) {
-            int newPosition = this.header.getHeaderLength()
-                    + this.header.getRecordLength() * (recno - 1);
+            long newPosition = this.header.getHeaderLength()
+                    + this.header.getRecordLength() * (long) (recno - 1);
 
             if (this.useMemoryMappedBuffer) {
-                buffer.position(newPosition);
+                buffer.position((int) newPosition);
             } else {
                 FileChannel fc = (FileChannel) this.channel;
                 fc.position(newPosition);
